Update/Fix
Hi Seth,
I wasn't able to save a KNN Model and the error appear to be at line 50 of this class.
I changed the format of the "ToString()" method from "r" to "d" because the parameter K is an integer and now it seems to work fine for my purpouse. Do you think it's correct?
I really appreciate your work. 
Many thanks,
Daniela